requirements.txt
fastapi==0.92.0
Pillow==9.4.0
python-dotenv==1.0.0
requests==2.27.1
